refactor logo e2e test code
