+++ visualization/app/codeCharta/e2e/logo.e2e.ts
@@ -1,25 +1,7 @@
 import {CC_URL, puppeteer} from "../../puppeteer.helper";
+import {LogoPageObject} from "./logo.po";
 
 jest.setTimeout(10000);
-
-export class LogoPageObject {
-
-    constructor(private page) {}
-
-    async getVersion() {
-        const versionString = await this.page.evaluate(() => document.querySelector('#title > div > h2 > span')["innerText"]);
-        return versionString.split(" ")[1];
-    }
-
-    async getLink() {
-        return await this.page.evaluate(() => document.querySelector('#title > div > a')["href"]);
-    }
-
-    async getImageSrc() {
-        return await this.page.evaluate(() => document.querySelector('#title > div > a > img')["src"]);
-    }
-
-}
 
 describe("CodeCharta logo tests",()=>{
